+++ crates/karva_cli/tests/cli.rs
@@ -11,142 +11,52 @@ use karva_core::{
     runner::{RunDiagnostics, TestRunner},
     testing::setup_module,
 };
-use karva_project::project::Project;
+use karva_project::{project::Project, tests::TestEnv};
 use rstest::rstest;
-use tempfile::TempDir;
 
 #[ctor::ctor]
 pub fn setup() {
     setup_module();
 }
 
-/// Find the karva wheel in the target/wheels directory.
-/// Returns the path to the wheel file.
-fn find_karva_wheel() -> anyhow::Result<PathBuf> {
-    let karva_root = Path::new(env!("CARGO_MANIFEST_DIR"))
-        .parent()
-        .and_then(|p| p.parent())
-        .ok_or_else(|| anyhow::anyhow!("Could not determine KARVA_ROOT"))?
-        .to_path_buf();
-
-    let wheels_dir = karva_root.join("target").join("wheels");
-
-    let entries = std::fs::read_dir(&wheels_dir)
-        .with_context(|| format!("Could not read wheels directory: {}", wheels_dir.display()))?;
-
-    for entry in entries {
-        let entry = entry?;
-        let file_name = entry.file_name();
-        if let Some(name) = file_name.to_str() {
-            if name.starts_with("karva-")
-                && std::path::Path::new(name)
-                    .extension()
-                    .is_some_and(|ext| ext.eq_ignore_ascii_case("whl"))
-            {
-                return Ok(entry.path());
-            }
-        }
-    }
-
-    anyhow::bail!("Could not find karva wheel in target/wheels directory");
-}
-
-struct TestCase {
-    _temp_dir: TempDir,
+struct IntegrationTestEnv {
+    test_env: TestEnv,
     _settings_scope: SettingsBindDropGuard,
-    project_dir: PathBuf,
 }
 
-impl TestCase {
-    fn new() -> anyhow::Result<Self> {
-        let temp_dir = TempDir::with_prefix("karva-test-env")?;
-
-        let project_dir = dunce::simplified(
-            &temp_dir
-                .path()
-                .canonicalize()
-                .context("Failed to canonicalize project path")?,
-        )
-        .to_path_buf();
-
-        let karva_wheel = find_karva_wheel()?;
-
-        let venv_path = project_dir.join(".venv");
-
-        // Set up a bare uv project and install pytest, mimicking the Python TestEnv
-        let commands = [
-            vec![
-                "uv",
-                "init",
-                "--bare",
-                "--directory",
-                project_dir.to_str().unwrap(),
-            ],
-            vec!["uv", "venv", venv_path.to_str().unwrap()],
-            vec![
-                "uv",
-                "pip",
-                "install",
-                "--python",
-                venv_path.to_str().unwrap(),
-                karva_wheel.to_str().unwrap(),
-                "pytest",
-            ],
-        ];
-
-        for command in &commands {
-            let output = Command::new(command[0])
-                .args(&command[1..])
-                .current_dir(&project_dir)
-                .output()
-                .with_context(|| format!("Failed to run command: {command:?}"))?;
-            if output.status.success() {
-                eprintln!(
-                    "Command succeeded: {:?}\nstdout:\n{}\nstderr:\n{}",
-                    command,
-                    String::from_utf8_lossy(&output.stdout),
-                    String::from_utf8_lossy(&output.stderr)
-                );
-            } else {
-                eprintln!(
-                    "Command failed: {:?}\nstdout:\n{}\nstderr:\n{}",
-                    command,
-                    String::from_utf8_lossy(&output.stdout),
-                    String::from_utf8_lossy(&output.stderr)
-                );
-                anyhow::bail!("Command failed: {:?}", command);
-            }
-        }
+impl IntegrationTestEnv {
+    fn new() -> Self {
+        let test_env = TestEnv::new();
 
         let mut settings = insta::Settings::clone_current();
-        settings.add_filter(&tempdir_filter(&project_dir), "<temp_dir>/");
+        settings.add_filter(&tempdir_filter(&test_env.cwd()), "<temp_dir>/");
         settings.add_filter(r#"\\(\w\w|\s|\.|")"#, "/$1");
 
         let settings_scope = settings.bind_to_scope();
 
-        Ok(Self {
-            project_dir,
-            _temp_dir: temp_dir,
+        Self {
+            test_env,
             _settings_scope: settings_scope,
-        })
+        }
     }
 
     fn karva_bin(&self) -> PathBuf {
         let venv_bin =
-            self.project_dir
+            self.test_env
+                .cwd()
                 .join(".venv")
                 .join(if cfg!(windows) { "Scripts" } else { "bin" });
         venv_bin.join(if cfg!(windows) { "karva.exe" } else { "karva" })
     }
 
     fn with_files<'a>(files: impl IntoIterator<Item = (&'a str, &'a str)>) -> anyhow::Result<Self> {
-        let case = Self::new()?;
+        let case = Self::new();
         case.write_files(files)?;
         Ok(case)
     }
 
     fn with_file(path: impl AsRef<Path>, content: &str) -> anyhow::Result<Self> {
-        let case = Self::new()?;
+        let case = Self::new();
         case.write_file(path, content)?;
         Ok(case)
     }
@@ -164,7 +74,7 @@ impl TestCase {
 
     fn write_file(&self, path: impl AsRef<Path>, content: &str) -> anyhow::Result<()> {
         let path = path.as_ref();
-        let path = self.project_dir.join(path);
+        let path = self.test_env.cwd().join(path);
 
         if let Some(parent) = path.parent() {
             std::fs::create_dir_all(parent)
@@ -178,7 +88,7 @@ impl TestCase {
 
     fn command(&self) -> Command {
         let mut command = Command::new(self.karva_bin());
-        command.current_dir(&self.project_dir).arg("test");
+        command.current_dir(self.test_env.cwd()).arg("test");
         command
     }
 
@@ -189,11 +99,11 @@ impl TestCase {
     }
 
     fn run(&self) -> RunDiagnostics {
-        self.run_with_path(&self.project_dir)
+        self.run_with_path(self.test_env.cwd())
     }
 
     fn run_with_path(&self, path: impl AsRef<Path>) -> RunDiagnostics {
-        let project = Project::new(self.project_dir.clone(), vec![path.as_ref().to_path_buf()]);
+        let project = Project::new(self.test_env.cwd(), vec![path.as_ref().to_path_buf()]);
         project.test_with_reporter(&mut DummyReporter)
     }
 }
@@ -279,7 +189,7 @@ fn tempdir_filter(path: &Path) -> String {
 
 #[test]
 fn test_one_test_passes() -> anyhow::Result<()> {
-    let case = TestCase::with_file(
+    let case = IntegrationTestEnv::with_file(
         "test_pass.py",
         r"
         def test_pass():
@@ -302,7 +212,7 @@ fn test_one_test_passes() -> anyhow::Result<()> {
 
 #[test]
 fn test_two_tests_pass() -> anyhow::Result<()> {
-    let case = TestCase::with_files([
+    let case = IntegrationTestEnv::with_files([
         (
             "test_pass.py",
             r"
@@ -335,7 +245,7 @@ fn test_two_tests_pass() -> anyhow::Result<()> {
 
 #[test]
 fn test_one_test_fails() -> anyhow::Result<()> {
-    let case = TestCase::with_file(
+    let case = IntegrationTestEnv::with_file(
         "test_fail.py",
         r"
         def test_fail():
@@ -361,7 +271,7 @@ fn test_one_test_fails() -> anyhow::Result<()> {
 
 #[test]
 fn test_multiple_tests_fail() -> anyhow::Result<()> {
-    let case = TestCase::with_file(
+    let case = IntegrationTestEnv::with_file(
         "test_fail2.py",
         r"
         def test_fail2():
@@ -387,7 +297,7 @@ fn test_multiple_tests_fail() -> anyhow::Result<()> {
 
 #[test]
 fn test_mixed_pass_and_fail() -> anyhow::Result<()> {
-    let case = TestCase::with_files([
+    let case = IntegrationTestEnv::with_files([
         (
             "test_pass.py",
             r"
@@ -423,7 +333,7 @@ fn test_mixed_pass_and_fail() -> anyhow::Result<()> {
 
 #[test]
 fn test_assertion_with_message() -> anyhow::Result<()> {
-    let case = TestCase::with_file(
+    let case = IntegrationTestEnv::with_file(
         "test_fail_with_msg.py",
         r#"
         def test_fail_with_message():
@@ -449,7 +359,7 @@ fn test_assertion_with_message() -> anyhow::Result<()> {
 
 #[test]
 fn test_equality_assertion_fail() -> anyhow::Result<()> {
-    let case = TestCase::with_file(
+    let case = IntegrationTestEnv::with_file(
         "test_equality.py",
         r"
         def test_equality():
@@ -477,7 +387,7 @@ fn test_equality_assertion_fail() -> anyhow::Result<()> {
 
 #[test]
 fn test_complex_assertion_fail() -> anyhow::Result<()> {
-    let case = TestCase::with_file(
+    let case = IntegrationTestEnv::with_file(
         "test_complex.py",
         r"
         def test_complex():
@@ -504,7 +414,7 @@ fn test_complex_assertion_fail() -> anyhow::Result<()> {
 
 #[test]
 fn test_long_file() -> anyhow::Result<()> {
-    let case = TestCase::with_file(
+    let case = IntegrationTestEnv::with_file(
         "test_long.py",
         r"
         # This is a long test file with many comments and functions
@@ -549,7 +459,7 @@ fn test_long_file() -> anyhow::Result<()> {
 
 #[test]
 fn test_multiple_assertions_in_function() -> anyhow::Result<()> {
-    let case = TestCase::with_file(
+    let case = IntegrationTestEnv::with_file(
         "test_multiple_assertions.py",
         r"
         def test_multiple_assertions():
@@ -578,7 +488,7 @@ fn test_multiple_assertions_in_function() -> anyhow::Result<()> {
 
 #[test]
 fn test_assertion_in_nested_function() -> anyhow::Result<()> {
-    let case = TestCase::with_file(
+    let case = IntegrationTestEnv::with_file(
         "test_nested.py",
         r"
         def helper():
@@ -608,7 +518,7 @@ fn test_assertion_in_nested_function() -> anyhow::Result<()> {
 
 #[test]
 fn test_assertion_with_complex_expression() -> anyhow::Result<()> {
-    let case = TestCase::with_file(
+    let case = IntegrationTestEnv::with_file(
         "test_complex_expr.py",
         r"
         def test_complex_expression():
@@ -635,7 +545,7 @@ fn test_assertion_with_complex_expression() -> anyhow::Result<()> {
 
 #[test]
 fn test_assertion_with_multiline_setup() -> anyhow::Result<()> {
-    let case = TestCase::with_file(
+    let case = IntegrationTestEnv::with_file(
         "test_multiline.py",
         r"
         def test_multiline_setup():
@@ -671,7 +581,7 @@ fn test_assertion_with_multiline_setup() -> anyhow::Result<()> {
 
 #[test]
 fn test_assertion_with_very_long_line() -> anyhow::Result<()> {
-    let case = TestCase::with_file(
+    let case = IntegrationTestEnv::with_file(
         "test_very_long_line.py",
         r"
         def test_very_long_line():
@@ -697,7 +607,7 @@ fn test_assertion_with_very_long_line() -> anyhow::Result<()> {
 
 #[test]
 fn test_assertion_on_line_1() -> anyhow::Result<()> {
-    let case = TestCase::with_file(
+    let case = IntegrationTestEnv::with_file(
         "test_line_1.py",
         r"def test_line_1():
     assert False",
@@ -721,7 +631,7 @@ fn test_assertion_on_line_1() -> anyhow::Result<()> {
 
 #[test]
 fn test_multiple_files_with_cross_function_calls() -> anyhow::Result<()> {
-    let case = TestCase::with_files([
+    let case = IntegrationTestEnv::with_files([
         (
             "helper.py",
             r"
@@ -762,7 +672,7 @@ fn test_multiple_files_with_cross_function_calls() -> anyhow::Result<()> {
 
 #[test]
 fn test_nested_function_calls_deep_stack() -> anyhow::Result<()> {
-    let case = TestCase::with_file(
+    let case = IntegrationTestEnv::with_file(
         "test_deep_stack.py",
         r"
         def level_1():
@@ -803,7 +713,7 @@ fn test_nested_function_calls_deep_stack() -> anyhow::Result<()> {
 
 #[test]
 fn test_assertion_in_class_method() -> anyhow::Result<()> {
-    let case = TestCase::with_file(
+    let case = IntegrationTestEnv::with_file(
         "test_class.py",
         r"
         class Calculator:
@@ -840,7 +750,7 @@ fn test_assertion_in_class_method() -> anyhow::Result<()> {
 
 #[test]
 fn test_assertion_in_imported_function() -> anyhow::Result<()> {
-    let case = TestCase::with_files([
+    let case = IntegrationTestEnv::with_files([
         (
             "validators.py",
             r"
@@ -933,7 +843,7 @@ fn test_fixture_initialization_order(#[case] framework: &str) -> anyhow::Result<
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(
         case,
@@ -987,7 +897,7 @@ fn test_empty_conftest() -> anyhow::Result<()> {
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
     success: true
@@ -1014,7 +924,7 @@ fn get_parametrize_function(package: &str) -> String {
 #[case("pytest")]
 #[case("karva")]
 fn test_parametrize(#[case] package: &str) -> anyhow::Result<()> {
-    let case = TestCase::with_file(
+    let case = IntegrationTestEnv::with_file(
         "test_parametrize.py",
         &format!(
             r"
@@ -1111,7 +1021,7 @@ fn test_function_scopes(#[case] framework: &str) -> anyhow::Result<()> {
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -1162,7 +1072,7 @@ fn test_module_scopes(#[case] framework: &str) -> anyhow::Result<()> {
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -1219,7 +1129,7 @@ fn test_package_scopes(#[case] framework: &str) -> anyhow::Result<()> {
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -1286,7 +1196,7 @@ fn test_session_scopes(#[case] framework: &str) -> anyhow::Result<()> {
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -1346,7 +1256,7 @@ fn test_mixed_scopes(#[case] framework: &str) -> anyhow::Result<()> {
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -1406,7 +1316,7 @@ fn test_fixture_across_files(#[case] framework: &str) -> anyhow::Result<()> {
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -1453,7 +1363,7 @@ fn test_named_fixtures(#[case] framework: &str) -> anyhow::Result<()> {
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -1518,7 +1428,7 @@ fn test_nested_package_scopes(#[case] framework: &str) -> anyhow::Result<()> {
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -1575,7 +1485,7 @@ fn test_independent_fixtures(#[case] framework: &str) -> anyhow::Result<()> {
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -1640,7 +1550,7 @@ fn test_multiple_files_independent_fixtures(#[case] framework: &str) -> anyhow::
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -1696,7 +1606,7 @@ fn test_basic_error_handling(#[case] framework: &str) -> anyhow::Result<()> {
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: false
@@ -1781,7 +1691,7 @@ fn test_different_scopes_independent(#[case] framework: &str) -> anyhow::Result<
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -1830,7 +1740,7 @@ fn test_invalid_scope_value(#[case] framework: &str) -> anyhow::Result<()> {
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: false
@@ -1881,7 +1791,7 @@ fn test_invalid_fixture_name(#[case] framework: &str) -> anyhow::Result<()> {
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: false
@@ -1945,7 +1855,7 @@ fn test_multiple_conftest_same_dir(#[case] framework: &str) -> anyhow::Result<()
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: false
@@ -2010,7 +1920,7 @@ fn test_very_deep_directory_structure(#[case] framework: &str) -> anyhow::Result
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -2059,7 +1969,7 @@ fn test_fixture_in_init_file(#[case] framework: &str) -> anyhow::Result<()> {
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: false
@@ -2129,7 +2039,7 @@ fn test_same_fixture_name_different_types(#[case] framework: &str) -> anyhow::Re
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, @r"
     success: true
@@ -2185,7 +2095,7 @@ fn test_fixture_dependencies(#[case] framework: &str) -> anyhow::Result<()> {
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -2271,7 +2181,7 @@ fn test_dependent_fixtures_different_scopes(#[case] framework: &str) -> anyhow::
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -2338,7 +2248,7 @@ fn test_complex_dependency_chain(#[case] framework: &str) -> anyhow::Result<()> 
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -2411,7 +2321,7 @@ fn test_mixed_scope_dependencies(#[case] framework: &str) -> anyhow::Result<()> 
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -2479,7 +2389,7 @@ fn test_diamond_dependency(#[case] framework: &str) -> anyhow::Result<()> {
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -2530,7 +2440,7 @@ fn test_generator_fixture(#[case] framework: &str) -> anyhow::Result<()> {
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -2591,7 +2501,7 @@ fn test_fixture_called_for_each_parametrization(#[case] framework: &str) -> anyh
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -2643,7 +2553,7 @@ fn test_fixture_finalizer_called_after_test(#[case] framework: &str) -> anyhow::
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -2699,7 +2609,7 @@ fn test_fixture_finalizer_called_at_correct_time(#[case] framework: &str) -> any
         ),
     ]);
 
-    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
 
     run_with_path_and_snapshot!(case, &["-s"], @r"
     success: true
@@ -2722,7 +2632,7 @@ fn test_fixture_finalizer_called_at_correct_time(#[case] framework: &str) -> any
 
 #[test]
 fn test_stdout_is_captured_and_displayed() -> anyhow::Result<()> {
-    let case = TestCase::with_file(
+    let case = IntegrationTestEnv::with_file(
         "test_std_out_redirected.py",
         r"
         def test_std_out_redirected():
@@ -2746,7 +2656,7 @@ fn test_stdout_is_captured_and_displayed() -> anyhow::Result<()> {
 
 #[test]
 fn test_stdout_is_captured_and_displayed_with_args() -> anyhow::Result<()> {
-    let case = TestCase::with_file(
+    let case = IntegrationTestEnv::with_file(
         "test_std_out_redirected.py",
         r"
         def test_std_out_redirected():
@@ -2806,7 +2716,7 @@ def test_string_and_int(order, first_entry):
 "#,
     );
 
-    let case = TestCase::with_file("test_append.py", &test_code)?;
+    let case = IntegrationTestEnv::with_file("test_append.py", &test_code)?;
 
     run_with_path_and_snapshot!(case,  @r"
     success: true
@@ -2846,7 +2756,7 @@ def test_value(clean_data):
 ",
     );
 
-    let case = TestCase::with_file("test_fixture_order_scope.py", &test_code)?;
+    let case = IntegrationTestEnv::with_file("test_fixture_order_scope.py", &test_code)?;
 
     run_with_path_and_snapshot!(case,  @r"
     success: true
